Moved onionshare and onionshare_gui logic directly into __init__.py files
